+++ android/src/main/kotlin/com/tundralabs/fluttertts/FlutterTtsPlugin.kt
@@ -175,7 +175,10 @@ class FlutterTtsPlugin : MethodCallHandler, FlutterPlugin {
 
     fun speakCompletion(success: Int) {
         speaking = false
-        handler!!.post { speakResult?.success(success) }
+        handler!!.post {
+            speakResult?.success(success)
+            speakResult = null
+        }
     }
 
     fun synthCompletion(success: Int) {
@@ -199,9 +202,12 @@ class FlutterTtsPlugin : MethodCallHandler, FlutterPlugin {
                 }
 
                 // Handle pending method calls (sent while TTS was initializing)
-                isTtsInitialized = true
-                for (call in pendingMethodCalls) {
-                    call.run()
+                synchronized(this@FlutterTtsPlugin) {
+                    isTtsInitialized = true
+                    for (call in pendingMethodCalls) {
+                        call.run()
+                    }
+                    pendingMethodCalls.clear()
                 }
                 invokeMethod("tts.init", isTtsInitialized)
             } else {
@@ -226,9 +232,12 @@ class FlutterTtsPlugin : MethodCallHandler, FlutterPlugin {
                 }
 
                 // Handle pending method calls (sent while TTS was initializing)
-                isTtsInitialized = true
-                for (call in pendingMethodCalls) {
-                    call.run()
+                synchronized(this@FlutterTtsPlugin) {
+                    isTtsInitialized = true
+                    for (call in pendingMethodCalls) {
+                        call.run()
+                    }
+                    pendingMethodCalls.clear()
                 }
             } else {
                 Log.e(tag, "Failed to initialize TextToSpeech with status: $status")
@@ -237,11 +246,13 @@ class FlutterTtsPlugin : MethodCallHandler, FlutterPlugin {
 
     override fun onMethodCall(call: MethodCall, result: Result) {
         // If TTS is still loading
-        if (!isTtsInitialized) {
-            // Suspend method call until the TTS engine is ready
-            val suspendedCall = Runnable { onMethodCall(call, result) }
-            pendingMethodCalls.add(suspendedCall)
-            return
+        synchronized(this@FlutterTtsPlugin) {
+            if (!isTtsInitialized) {
+                // Suspend method call until the TTS engine is ready
+                val suspendedCall = Runnable { onMethodCall(call, result) }
+                pendingMethodCalls.add(suspendedCall)
+                return
+            }
         }
         when (call.method) {
             "speak" -> {
@@ -269,8 +280,10 @@ class FlutterTtsPlugin : MethodCallHandler, FlutterPlugin {
                 }
                 val b = speak(text)
                 if (!b) {
-                    val suspendedCall = Runnable { onMethodCall(call, result) }
-                    pendingMethodCalls.add(suspendedCall)
+                    synchronized(this@FlutterTtsPlugin) {
+                        val suspendedCall = Runnable { onMethodCall(call, result) }
+                        pendingMethodCalls.add(suspendedCall)
+                    }
                     return
                 }
                 // Only use await speak completion if queueMode is set to QUEUE_FLUSH