+++ lib/flutter_tts.dart
@@ -507,6 +507,7 @@ class FlutterTts {
 
   /// [Future] which invokes the platform specific method for getVoices
   /// Returns a `List` of `Maps` containing a voice name and locale
+  /// For iOS specifically, it also includes quality, gender, and identifier
   /// ***Android, iOS, and macOS supported only***
   Future<dynamic> get getVoices async {
     final voices = await _channel.invokeMethod('getVoices');